chore: more clippy
Very weird that this one doesn't show up locally without nightly - CI doesn't seem to be using nightly, but this fails there

+++ crates/owl-recorder/src/recorder.rs
@@ -295,7 +295,7 @@ pub fn get_recording_base_resolution(hwnd: HWND) -> Result<(u32, u32)> {
         .ok()
         .ok()?;
 
-        Some((devmode.dmPelsWidth as u32, devmode.dmPelsHeight as u32))
+        Some((devmode.dmPelsWidth, devmode.dmPelsHeight))
     }
 
     match get_window_inner_size(hwnd) {
